requirements.txt
# Deep Learning Framework
torch==2.1.0
torchvision==0.16.0

# Data Processing
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0

# Image Processing
Pillow==10.0.0
opencv-python==4.8.0.76

# Visualization
matplotlib==3.7.2
seaborn==0.12.2

# Progress Bars
tqdm==4.66.1

# Jupyter Notebook
jupyter==1.0.0
ipykernel==6.25.1

# Model Evaluation
scikit-image==0.21.0

# Dataset Download (Kaggle)
kaggle==1.5.16